--- thoth/storages/data/alembic/versions/f6d46e0b37da_flag_for_if_specific_hash_has_gone_.py
@@ -1,28 +1,0 @@
-"""Flag for if specific hash has gone missing, defaults to present
-
-Revision ID: f6d46e0b37da
-Revises: 74254b596f00
-Create Date: 2020-03-24 21:32:57.426880+00:00
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'f6d46e0b37da'
-down_revision = '74254b596f00'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('python_artifact', sa.Column('present', sa.Boolean(), nullable=False, server_default="false"))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('python_artifact', 'present')
-    # ### end Alembic commands ###
